menambahkan get product dan tampilan create product

+++ app/src/main/java/com/workshopkotlin/anggitprayogo/data/entity/UserEntity.kt
@@ -7,7 +7,7 @@ import android.arch.persistence.room.PrimaryKey
 
 @Entity(tableName = "users", indices = [Index(value = ["email"], unique = true)])
 data class UserEntity(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id_user") var id: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id_user") var id: Long = 0,
     @ColumnInfo(name = "name") var name: String = "",
     @ColumnInfo(name = "email") var email: String = ""
 )
